move weather functions to favorites
